// web/admin/src/types/api/v1/models/AuthProvider.ts
/* tslint:disable */
/* eslint-disable */
/**
 * ふるマル API - 管理者向け
 * 管理者向けのふるマルAPIです。
 *
 * The version of the OpenAPI document: 
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { AuthProviderType } from './AuthProviderType';
import {
    AuthProviderTypeFromJSON,
    AuthProviderTypeFromJSONTyped,
    AuthProviderTypeToJSON,
    AuthProviderTypeToJSONTyped,
} from './AuthProviderType';

/**
 * 
 * @export
 * @interface AuthProvider
 */
export interface AuthProvider {
    /**
     * 連携日時
     * @type {number}
     * @memberof AuthProvider
     */
    connectedAt: number;
    /**
     * 
     * @type {AuthProviderType}
     * @memberof AuthProvider
     */
    type: AuthProviderType;
}



/**
 * Check if a given object implements the AuthProvider interface.
 */
export function instanceOfAuthProvider(value: object): value is AuthProvider {
    if (!('connectedAt' in value) || value['connectedAt'] === undefined) return false;
    if (!('type' in value) || value['type'] === undefined) return false;
    return true;
}

export function AuthProviderFromJSON(json: any): AuthProvider {
    return AuthProviderFromJSONTyped(json, false);
}

export function AuthProviderFromJSONTyped(json: any, ignoreDiscriminator: boolean): AuthProvider {
    if (json == null) {
        return json;
    }
    return {
        
        'connectedAt': json['connectedAt'],
        'type': AuthProviderTypeFromJSON(json['type']),
    };
}

export function AuthProviderToJSON(json: any): AuthProvider {
    return AuthProviderToJSONTyped(json, false);
}

export function AuthProviderToJSONTyped(value?: AuthProvider | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'connectedAt': value['connectedAt'],
        'type': AuthProviderTypeToJSON(value['type']),
    };
}