// web/admin/src/types/api/v1/models/AuthResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * ふるマル API - 管理者向け
 * 管理者向けのふるマルAPIです。
 *
 * The version of the OpenAPI document: 
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { AdminType } from './AdminType';
import {
    AdminTypeFromJSON,
    AdminTypeFromJSONTyped,
    AdminTypeToJSON,
    AdminTypeToJSONTyped,
} from './AdminType';

/**
 * 
 * @export
 * @interface AuthResponse
 */
export interface AuthResponse {
    /**
     * アクセストークン
     * @type {string}
     * @memberof AuthResponse
     */
    accessToken: string;
    /**
     * 管理者ID
     * @type {string}
     * @memberof AuthResponse
     */
    adminId: string;
    /**
     * 有効期限
     * @type {number}
     * @memberof AuthResponse
     */
    expiresIn: number;
    /**
     * 更新トークン
     * @type {string}
     * @memberof AuthResponse
     */
    refreshToken: string;
    /**
     * トークン種別
     * @type {string}
     * @memberof AuthResponse
     */
    tokenType: string;
    /**
     * 
     * @type {AdminType}
     * @memberof AuthResponse
     */
    type: AdminType;
}



/**
 * Check if a given object implements the AuthResponse interface.
 */
export function instanceOfAuthResponse(value: object): value is AuthResponse {
    if (!('accessToken' in value) || value['accessToken'] === undefined) return false;
    if (!('adminId' in value) || value['adminId'] === undefined) return false;
    if (!('expiresIn' in value) || value['expiresIn'] === undefined) return false;
    if (!('refreshToken' in value) || value['refreshToken'] === undefined) return false;
    if (!('tokenType' in value) || value['tokenType'] === undefined) return false;
    if (!('type' in value) || value['type'] === undefined) return false;
    return true;
}

export function AuthResponseFromJSON(json: any): AuthResponse {
    return AuthResponseFromJSONTyped(json, false);
}

export function AuthResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): AuthResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'accessToken': json['accessToken'],
        'adminId': json['adminId'],
        'expiresIn': json['expiresIn'],
        'refreshToken': json['refreshToken'],
        'tokenType': json['tokenType'],
        'type': AdminTypeFromJSON(json['type']),
    };
}

export function AuthResponseToJSON(json: any): AuthResponse {
    return AuthResponseToJSONTyped(json, false);
}

export function AuthResponseToJSONTyped(value?: AuthResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'accessToken': value['accessToken'],
        'adminId': value['adminId'],
        'expiresIn': value['expiresIn'],
        'refreshToken': value['refreshToken'],
        'tokenType': value['tokenType'],
        'type': AdminTypeToJSON(value['type']),
    };
}